requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
apscheduler==3.10.1
requests==2.31.0